% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_ccfrp.R
\name{clean_ccfrp}
\alias{clean_ccfrp}
\title{Clean and standardize CCFRP data}
\usage{
clean_ccfrp(dir, species, data)
}
\arguments{
\item{dir}{Directory location to save the cleaned data frame}

\item{species}{add definition}
}
\description{
Clean and standardize CCFRP data
}
\author{
Chantel Wetzel
}
