% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_swfsc_cpfv.R
\name{clean_coop_samples}
\alias{clean_coop_samples}
\title{Clean COOP data from Melissa Monk and Rachel Brooks}
\usage{
clean_coop_samples(data, species)
}
\arguments{
\item{data}{add definition}

\item{species}{add definition}
}
\description{
Clean COOP data from Melissa Monk and Rachel Brooks
}
\author{
Chantel Wetzel
}
