% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_nwfsc_hkl.R
\name{clean_nwfsc_hkl}
\alias{clean_nwfsc_hkl}
\title{Function to remove all unneeded species data and to format column names as needed for the NWFS HKL data.}
\usage{
clean_nwfsc_hkl(dir = here::here("data-processed"), data, species)
}
\arguments{
\item{dir}{Directory location to save the cleaned data frame}

\item{data}{Data frame of NWFSC HKL data}

\item{species}{A list of species names created by the get_species_list function}
}
\description{
Function to remove all unneeded species data and to format column names as needed for the NWFS HKL data.
}
\author{
Chantel Wetzel
}
