% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_pacfin_comps.R
\name{clean_pacfin_comps}
\alias{clean_pacfin_comps}
\title{Clean pulled PacFIN data}
\usage{
clean_pacfin_comps(dir, bds_pacfin, species, spid_key, year = 1980)
}
\arguments{
\item{dir}{Directory location to save the cleaned data frame}

\item{bds_pacfin}{Dataframe of PacFIN data}

\item{species}{add definition}

\item{spid_key}{add definition}

\item{year}{Integer to filter PacFIN data to retain all data from this year
and beyond}
}
\description{
Clean pulled PacFIN data
}
\author{
Chantel Wetzel
}
