% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_wcgbt_bio.R
\name{clean_wcgbt_bio}
\alias{clean_wcgbt_bio}
\title{Clean NWFSC WCGBTS biological data}
\usage{
clean_wcgbt_bio(dir = here::here("data-processed"), species, data)
}
\arguments{
\item{dir}{Directory location to save the cleaned data frame}

\item{species}{A list of species names created by the get_species_list function}

\item{data}{List of catch and bio data to clean up}
}
\description{
Clean NWFSC WCGBTS biological data
}
\author{
Chantel Wetzel
}
