% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_wcgbt_catch.R
\name{clean_wcgbt_catch}
\alias{clean_wcgbt_catch}
\title{Clean NWFSC WCGBTS survey catch file}
\usage{
clean_wcgbt_catch(dir = here::here("data-raw"), species, data)
}
\arguments{
\item{dir}{Directory location to save the cleaned data frame}

\item{species}{A list of species names created by the get_species_list function}

\item{data}{List of catch and bio data to clean up}
}
\description{
Clean NWFSC WCGBTS survey catch file
}
\author{
Chantel Wetzel
}
