% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_species.R
\name{combine_species}
\alias{combine_species}
\title{Helper function to standardize species names}
\usage{
combine_species(data, name = "Common_name")
}
\arguments{
\item{data}{add definition}

\item{name}{Common_name}
}
\description{
Helper function to standardize species names
}
\author{
Chantel Wetzel
}
