% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_cdfw_otolith_files.R
\name{format_cdfw_otolith_files}
\alias{format_cdfw_otolith_files}
\title{Format otolith data}
\usage{
format_cdfw_otolith_files(data, source = NA, state = NA)
}
\arguments{
\item{data}{add definition}

\item{source}{add definition}

\item{state}{add definition}
}
\description{
Format otolith data
}
\author{
Chantel Wetzel
}
