% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_hkl_species.R
\name{get_hkl_species}
\alias{get_hkl_species}
\title{List of species name to estimate indices for NWFSC HKL survey data}
\usage{
get_hkl_species()
}
\description{
List of species name to estimate indices for NWFSC HKL survey data
}
\author{
Chantel Wetzel
}
