% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_species_list.R
\name{get_species_list}
\alias{get_species_list}
\title{List of species names to pull from NWFSC survey data}
\usage{
get_species_list()
}
\description{
List of species names to pull from NWFSC survey data
}
\author{
Chantel Wetzel
}
