% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_pacfin_data.R
\name{load_pacfin_data}
\alias{load_pacfin_data}
\title{Function to load in a PacFIN Rdata file.}
\usage{
load_pacfin_data(dir, file_name)
}
\arguments{
\item{dir}{Directory location to save the cleaned data frame}

\item{file_name}{PacFIN Rdata file to be loaded}
}
\description{
The function expects the Rdata file to have a name of bds.pacfin.
}
\author{
Chantel Wetzel
}
