% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_wcgbt_comps.R
\name{plot_wcgbt_comps}
\alias{plot_wcgbt_comps}
\title{Plot NWFSC WCGBTS composition data}
\usage{
plot_wcgbt_comps(dir = here::here(), wcgbt_catch, wcgbt_bio)
}
\arguments{
\item{dir}{add definition}

\item{wcgbt_catch}{add definition}

\item{wcgbt_bio}{add definition}
}
\description{
Plot NWFSC WCGBTS composition data
}
\author{
Chantel Wetzel
}
