% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_unexpanded_comps.R
\name{process_comps_unexpanded}
\alias{process_comps_unexpanded}
\title{Creates a matrix of length or age composition data WITHOUT expansion}
\usage{
process_comps_unexpanded(
  dir = NULL,
  datL,
  lgthBins = 1,
  sex = lifecycle::deprecated(),
  partition = 0,
  fleet = "Enter Fleet",
  ageErr = "NA",
  agelow = -1,
  agehigh = -1,
  month = "Enter Month",
  two_sex_model = TRUE,
  printfolder = "forSS",
  verbose = TRUE
)
}
\arguments{
\item{dir}{directory this is where the output files will be saved}

\item{datL}{the read in length comps by the PullBio.fn function}

\item{lgthBins}{A vector on length bins or age bins to group the data for Stock Synthsis}

\item{sex}{Deprecated with {nwfscSurvey} 2.2.1 (February 2023). The function will now output
female/male and unsexed compositions if present in the data frame datL}

\item{partition}{partition for Stock Synthesis}

\item{fleet}{fleet number}

\item{ageErr}{Number of ageing error matrix for Stock Synthesis}

\item{agelow}{age bin for Stock Synthesis (default value of -1)}

\item{agehigh}{age bin for Stock Synthesis (default value of -1)}

\item{month}{month the samples were collected used by Stock Synthesis to determine the length/age estimate to compare to.}

\item{two_sex_model}{Default TRUE. If TRUE and unsexed composition data are present the unsexed comps
will be output in the format needed for a two-sex model in Stock Synthesis.}

\item{printfolder}{folder where the length comps will be saved}

\item{verbose}{A logical value specifying whether or not to print out
message statements to the screen while using this function.
The default is to print all messages, i.e., \code{verbose = TRUE}.}
}
\description{
Creates a matrix of length or age composition data WITHOUT expansion
}
\author{
Chantel Wetzel
}
