% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_wcgbts.R
\name{pull_wcgbts}
\alias{pull_wcgbts}
\title{Wrapper function to pull all species catch and biological data from the NWFSC data wharehouse.}
\usage{
pull_wcgbts(
  dir = here::here("data-raw"),
  species,
  survey = "NWFSC.Combo",
  load = FALSE,
  convert = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{dir}{Directory location to save the cleaned data frame}

\item{species}{A list of species names created by the get_species_list function}

\item{survey}{A character entry from one of the following options that
specifies which survey to pull the data for:
\itemize{
\item Triennial,
\item AFSC.Slope,
\item NWFSC.Combo,
\item NWFSC.Slope,
\item NWFSC.Shelf
}}

\item{load}{Option to load in existing Rdata files with the catch and biological data.}

\item{convert}{A logical as to whether to convert the data column names
to the old names as currently used in the nwfscSurvey processing code}

\item{verbose}{A logical value specifying whether or not to print out
message statements to the screen while using this function.
The default is to print all messages, i.e., \code{verbose = TRUE}.}
}
\description{
Wrapper function to pull all species catch and biological data from the NWFSC data wharehouse.
}
\author{
Chantel Wetzel
}
