% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_new_survey_information.R
\name{summarize_survey_new_information}
\alias{summarize_survey_new_information}
\title{Summarize new survey data available for a new assessment}
\usage{
summarize_survey_new_information(dir, stock_year, wcgbt, hkl)
}
\arguments{
\item{dir}{add description}

\item{stock_year}{add description}

\item{wcgbt}{description}

\item{hkl}{add description}
}
\description{
Summarize new survey data available for a new assessment
}
\author{
Chantel Wetzel
}
